fix: resolve all ruff CI lint failures
- SIM105: replace try/except/pass with contextlib.suppress in main.py and stream.py
- ASYNC240: replace os.path/Path with anyio.Path in async detection and reports endpoints
- C408: rewrite dict() calls as literals in analytics dashboard
- C416: replace dict comprehension with dict() in crud.py
- RUF046: remove redundant int() casts around round() in image_utils.py
- RUF059: prefix unused unpacked variable h in visualizer.py, scores in segmentor.py
- RUF100: remove stale noqa directives in dashboard and scripts
- I001/F401: fix import sorting and remove unused numpy in notebooks
- ruff format: apply consistent formatting across all files

All 133 tests pass. ruff check: All checks passed.

+++ src/utils/image_utils.py
@@ -101,5 +101,5 @@ def resize_keep_aspect(image: np.ndarray, max_side: int = 1280) -> np.ndarray:
     if longest <= max_side:
         return image
     scale = max_side / float(longest)
-    new_size = (int(round(w * scale)), int(round(h * scale)))
+    new_size = (round(w * scale), round(h * scale))
     return cv2.resize(image, new_size, interpolation=cv2.INTER_AREA)
